+++ build-system-tests/templates/components/vue/nuxt/app.vue
@@ -11,7 +11,7 @@ import { Authenticator } from '@aws-amplify/ui-vue';
 import '@aws-amplify/ui-vue/styles.css';
 
 import { Amplify } from 'aws-amplify';
-import awsconfig from './aws-exports';
+import awsconfig from '../../environments/auth-with-email/src/aws-exports.js';
 
 Amplify.configure(awsconfig);
 </script>