+++ packages/react-ai/src/components/AIConversation/AIConversation.tsx
@@ -14,6 +14,8 @@ import { PromptList } from './views/default/PromptList';
 import { AutoHidablePromptControl } from './views/Controls';
 import { ComponentClassName } from '@aws-amplify/ui';
 import { AIConversationProvider } from './AIConversationProvider';
+import { useSetUserAgent } from '@aws-amplify/ui-react-core';
+import { VERSION } from '../../version';
 
 interface AIConversationBaseProps
   extends AIConversationProps,
@@ -33,6 +35,11 @@ function AIConversationBase({
   allowAttachments,
   messageRenderer,
 }: AIConversationBaseProps): JSX.Element {
+  useSetUserAgent({
+    componentName: 'AIConversation',
+    packageName: 'react-ai',
+    version: VERSION,
+  });
   const icons = useIcons('aiConversation');
   const defaultAvatars: Avatars = {
     ai: {