+++ packages/react-native-auth/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aws-amplify/ui-react-native-auth
 
+## 0.0.33
+
+### Patch Changes
+
+- Updated dependencies [[`ac7cb271a`](https://github.com/aws-amplify/amplify-ui/commit/ac7cb271aff895e643fb5dc927030df9245b7c5b)]:
+  - @aws-amplify/ui@6.6.4
+  - @aws-amplify/ui-react-core-auth@0.0.30
+  - @aws-amplify/ui-react-native@2.2.15
+
 ## 0.0.32
 
 ### Patch Changes